Move test from bootstrap to compiletest
