--- src/test/rustdoc-test/test.rs
@@ -1,14 +1,0 @@
-/// This is a Foo;
-///
-/// ```
-/// println!("baaaaaar");
-/// ```
-#[unstable]
-pub struct Foo;
-
-/// This is a Bar;
-///
-/// ```
-/// println!("fooooo");
-/// ```
-pub struct Bar;
